+++ src/golang/cmd/server/handler/get_artifact_versions.go
@@ -2,13 +2,13 @@ package handler
 
 import (
 	"context"
+	"github.com/aqueducthq/aqueduct/lib/collections/operator/connector"
 	"net/http"
 
 	"github.com/aqueducthq/aqueduct/cmd/server/queries"
 	"github.com/aqueducthq/aqueduct/lib/collections/shared"
 	aq_context "github.com/aqueducthq/aqueduct/lib/context"
 	"github.com/aqueducthq/aqueduct/lib/database"
-	"github.com/aqueducthq/aqueduct/lib/workflow/operator/connector"
 	"github.com/dropbox/godropbox/errors"
 	"github.com/google/uuid"
 )
