+++ src/golang/lib/collections/operator/connector/extract.go
@@ -2,7 +2,6 @@ package connector
 
 import (
 	"encoding/json"
-
 	"github.com/aqueducthq/aqueduct/lib/collections/integration"
 	"github.com/dropbox/godropbox/errors"
 	"github.com/google/uuid"
