+++ src/golang/lib/collections/operator/connector/extract_test.go
