+++ src/golang/lib/collections/operator/connector/github/types.go
@@ -1,4 +1,4 @@
-package types
+package github
 
 import (
 	"fmt"
