+++ src/golang/lib/collections/operator/connector/load_params.go
