+++ src/golang/lib/collections/operator/connector/load_test.go
