+++ src/golang/lib/collections/operator/metric/metric.go
@@ -1,7 +1,7 @@
 package metric
 
 import (
-	"github.com/aqueducthq/aqueduct/lib/workflow/operator/function"
+	"github.com/aqueducthq/aqueduct/lib/collections/operator/function"
 )
 
 type Metric struct {
