+++ src/golang/lib/collections/operator/param/param.go
