+++ src/golang/lib/collections/tests/operator_test.go
@@ -2,12 +2,12 @@ package tests
 
 import (
 	"context"
+	connector2 "github.com/aqueducthq/aqueduct/lib/collections/operator/connector"
+	"github.com/aqueducthq/aqueduct/lib/collections/operator/function"
 	"testing"
 
 	"github.com/aqueducthq/aqueduct/lib/collections/integration"
 	"github.com/aqueducthq/aqueduct/lib/collections/operator"
-	"github.com/aqueducthq/aqueduct/lib/workflow/operator/connector"
-	"github.com/aqueducthq/aqueduct/lib/workflow/operator/function"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 )
@@ -62,11 +62,11 @@ func TestCreateOperator(t *testing.T) {
 	expectedOperator := &operator.DBOperator{
 		Name:        "test-operator",
 		Description: "testing op",
-		Spec: *operator.NewSpecFromExtract(connector.Extract{
+		Spec: *operator.NewSpecFromExtract(connector2.Extract{
 			Service:       integration.Postgres,
 			IntegrationId: integrations[0].Id,
-			Parameters: &connector.PostgresExtractParams{
-				connector.RelationalDBExtractParams{
+			Parameters: &connector2.PostgresExtractParams{
+				connector2.RelationalDBExtractParams{
 					Query: "SELECT * FROM mpg;",
 				},
 			},
