+++ src/golang/lib/workflow/scheduler/load.go
@@ -3,11 +3,11 @@ package scheduler
 import (
 	"context"
 	"fmt"
+	"github.com/aqueducthq/aqueduct/lib/collections/operator/connector"
 
 	"github.com/aqueducthq/aqueduct/lib/collections/shared"
 	"github.com/aqueducthq/aqueduct/lib/job"
 	"github.com/aqueducthq/aqueduct/lib/vault"
-	"github.com/aqueducthq/aqueduct/lib/workflow/operator/connector"
 	"github.com/aqueducthq/aqueduct/lib/workflow/operator/connector/auth"
 	"github.com/dropbox/godropbox/errors"
 	"github.com/google/uuid"
